#9 Support cancel notifications.
This is likely to break some applications - basically, if you're consuming from a queue that is deleted, you will receive 'basic.cancel' instead of a valid message. This is a good solution though - it's better to crash user app (that is not expecting this behaviour) than to allow it to hang indefinetely.