Use bit_length to determine encoding of the integer
Previous implementation was wrong, not only the binary representaiton could start with `0b` instead of `-0b`, but also int type on 64 bit machines will be 64 bit long.

Additionally, I don't think that encoding long integers as strings is a good idea. Let's throw an error instead and let the programmer deal with this.